improve code structure fix  bug
