+++ src/Pipboy.Avalonia.Fx/AssemblyInfo.cs
@@ -2,4 +2,4 @@
 
 [assembly: XmlnsPrefix("https://github.com/NeverMorewd/Pipboy.Avalonia", "pb")]
 [assembly: XmlnsDefinition("https://github.com/NeverMorewd/Pipboy.Avalonia", "Pipboy.Avalonia.Fx")]
-[assembly: XmlnsDefinition("https://github.com/NeverMorewd/Pipboy.Avalonia", "PipBoy.Avalonia.Fx.Controls")]
+[assembly: XmlnsDefinition("https://github.com/NeverMorewd/Pipboy.Avalonia", "Pipboy.Avalonia.Fx.Controls")]
